Move DirectDraw code to its own .h and .cpp files.
