+++ desmume/src/frontend/windows/DeSmuME.vcxproj
@@ -284,6 +284,7 @@
     <ClCompile Include="colorctrl.cpp" />
     <ClCompile Include="console.cpp" />
     <ClCompile Include="CWindow.cpp" />
+    <ClCompile Include="ddraw.cpp" />
     <ClCompile Include="FEX_Interface.cpp" />
     <ClCompile Include="File_Extractor\7z_C\7zAlloc.c" />
     <ClCompile Include="File_Extractor\7z_C\7zBuf.c" />
@@ -582,6 +583,7 @@
     <ClInclude Include="colorctrl.h" />
     <ClInclude Include="console.h" />
     <ClInclude Include="CWindow.h" />
+    <ClInclude Include="ddraw.h" />
     <ClInclude Include="FEX_Interface.h" />
     <ClInclude Include="File_Extractor\7z_C\7z.h" />
     <ClInclude Include="File_Extractor\7z_C\7zAlloc.h" />
