+++ desmume/src/frontend/windows/DeSmuME.vcxproj.filters
@@ -921,6 +921,9 @@
     <ClCompile Include="..\..\utils\colorspacehandler\colorspacehandler_AltiVec.cpp">
       <Filter>utils\colorspacehandler</Filter>
     </ClCompile>
+    <ClCompile Include="ddraw.cpp">
+      <Filter>frontend\Windows</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\utils\guid.h">
@@ -1609,6 +1612,9 @@
     </ClInclude>
     <ClInclude Include="..\..\utils\colorspacehandler\colorspacehandler_SSE2.h">
       <Filter>utils\colorspacehandler</Filter>
+    </ClInclude>
+    <ClInclude Include="ddraw.h">
+      <Filter>frontend\Windows</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
